Embedding serialization

+++ src/main/java/ldr/client/domen/db/DataBase.java
@@ -20,7 +20,7 @@ public class DataBase implements IDataBase {
     }
 
     @Override
-    public void createCollection(String name, long dimension, Metric metric) throws DataBaseException {
+    public void createCollection(String name, int dimension, Metric metric) throws DataBaseException {
         // Проверка на то, что такой коллекции нет, если есть выкидываем exception
 
         // В файловой системе создается новая папка с таким name
